pyproject.toml
[tool.poetry]
name = "aifu"
version = "0.1.0"
description = "AIFU is Rag for asking informations related to AI such as waifu tier list based on user preferences, recommended waifus for the user, or we can se another users waifu tier lists that are they saved"
authors = ["abdurrahmanharitsghiffary <abdmanharits@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.9"
pre-commit = "^4.0.1"
langchain-google-vertexai = "^2.0.8"
langchain-google-genai = "^2.0.6"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
pypdf = "^5.1.0"
langchain-community = "^0.3.8"
langchain-chroma = "^0.1.4"


[tool.poetry.group.dev.dependencies]
commitizen = "^4.0.0"
langchain-cli = "^0.0.33"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
